serverless/lambda/requirements.txt
bcrypt==4.1.3
pyotp==2.9.0
PyJWT==2.8.0
